restructure example files, update docs and build
